+++ app/page.tsx
@@ -1,16 +1,15 @@
 
+import MusicPlayer from "@/components/border-beam-demo-2"
 import Features from "@/components/features-1"
 import { HeroHeader } from "@/components/header"
-import HeroSection from "@/components/hero-section"
-import { LandingPage } from "@/components/landing-page"
+import HeroSection from "@/components/lending-page"
+import { Music } from "lucide-react"
 
 export default function Home() {
   return (
     <>
     
       <HeroSection />
-      {/* <Features /> */}
-      <LandingPage />
     </>
   )
