+++ components/footer.tsx
@@ -1,0 +1,120 @@
+import { Github, Twitter, MessageCircle, Lock } from 'lucide-react';
+import React from 'react';
+import { cn } from '@/lib/utils';
+
+export default function Footer() {
+    return (
+        <>
+            {/* Footer */}
+            <footer className="bg-gray-900 text-white py-16 px-4 sm:px-6 lg:px-8">
+                <div className="max-w-7xl mx-auto">
+                    <div className="grid md:grid-cols-4 gap-8 mb-12">
+                        <div className="md:col-span-2">
+                            <div className="flex items-center space-x-3 mb-6">
+                                <div className="w-12 h-12 bg-gradient-to-r from-blue-600 to-purple-600 rounded-xl flex items-center justify-center">
+                                    <span className="text-white font-bold text-lg">M</span>
+                                </div>
+                                <div>
+                                    <span className="text-2xl font-bold">MomentumFi</span>
+                                    <p className="text-sm text-gray-400">AI-Powered DeFi</p>
+                                </div>
+                            </div>
+                            <p className="text-gray-400 mb-6 max-w-md">
+                                The future of decentralized finance powered by artificial intelligence and Internet Computer Protocol.
+                            </p>
+                            <div className="flex items-center space-x-4">
+                                <a
+                                    href="#"
+                                    className="w-10 h-10 bg-gray-800 rounded-lg flex items-center justify-center hover:bg-gray-700 transition-colors"
+                                >
+                                    <Github className="h-5 w-5" />
+                                </a>
+                                <a
+                                    href="#"
+                                    className="w-10 h-10 bg-gray-800 rounded-lg flex items-center justify-center hover:bg-gray-700 transition-colors"
+                                >
+                                    <Twitter className="h-5 w-5" />
+                                </a>
+                                <a
+                                    href="#"
+                                    className="w-10 h-10 bg-gray-800 rounded-lg flex items-center justify-center hover:bg-gray-700 transition-colors"
+                                >
+                                    <MessageCircle className="h-5 w-5" />
+                                </a>
+                            </div>
+                        </div>
+
+                        <div>
+                            <h3 className="text-lg font-semibold mb-4">Product</h3>
+                            <ul className="space-y-2 text-gray-400">
+                                <li>
+                                    <a href="#" className="hover:text-white transition-colors">
+                                        Dashboard
+                                    </a>
+                                </li>
+                                <li>
+                                    <a href="#" className="hover:text-white transition-colors">
+                                        Analytics
+                                    </a>
+                                </li>
+                                <li>
+                                    <a href="#" className="hover:text-white transition-colors">
+                                        API
+                                    </a>
+                                </li>
+                                <li>
+                                    <a href="#" className="hover:text-white transition-colors">
+                                        Documentation
+                                    </a>
+                                </li>
+                            </ul>
+                        </div>
+
+                        <div>
+                            <h3 className="text-lg font-semibold mb-4">Resources</h3>
+                            <ul className="space-y-2 text-gray-400">
+                                <li>
+                                    <a href="#" className="hover:text-white transition-colors">
+                                        Blog
+                                    </a>
+                                </li>
+                                <li>
+                                    <a href="#" className="hover:text-white transition-colors">
+                                        Help Center
+                                    </a>
+                                </li>
+                                <li>
+                                    <a href="#" className="hover:text-white transition-colors">
+                                        Community
+                                    </a>
+                                </li>
+                                <li>
+                                    <a href="#" className="hover:text-white transition-colors">
+                                        Status
+                                    </a>
+                                </li>
+                            </ul>
+                        </div>
+                    </div>
+
+                    <div className="border-t border-gray-800 pt-8">
+                        <div className="flex flex-col md:flex-row justify-between items-center">
+                            <p className="text-gray-400 text-sm mb-4 md:mb-0">© 2024 MomentumFi. All rights reserved.</p>
+                            <div className="flex items-center space-x-6">
+                                <p className="text-yellow-400 text-sm font-semibold flex items-center">
+                                    <Lock className="h-4 w-4 mr-2" />
+                                    Built for Hackathon Demo Purposes
+                                </p>
+                                <a href="#" className="text-gray-400 hover:text-white text-sm transition-colors">
+                                    Privacy Policy
+                                </a>
+                                <a href="#" className="text-gray-400 hover:text-white text-sm transition-colors">
+                                    Terms of Service
+                                </a>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </footer></>
+    )
+}
